refactor(engine): ♻️  separation of responsibilities
