+++ packages/engine/src/index.ts
@@ -1,1 +1,1 @@
-export { default as Engine } from "./Engine"
+export { default as Engine } from "./engine"
